+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add a `HtmlPruner` class
+  ([#679](https://github.com/MyIntervals/emogrifier/pull/679))
 - Add `AbstractHtmlProcessor::fromDomDocument`
   ([#676](https://github.com/MyIntervals/emogrifier/pull/676))
 - Add `AbstractHtmlProcessor::fromHtml`